+++ commands/heist.js
@@ -11,7 +11,7 @@ var timeout
 var players = []
 var reward
 
-var timer = 30000
+var timer = 60000
 var playervalue = rn({min: 0.01, max: 0.2})
 
 exports.run = (message, bot) => {
@@ -29,7 +29,7 @@ exports.run = (message, bot) => {
             if(rand < winrate) {
                 reward = rn({min: settings.actions.heist.minwin, max: settings.actions.heist.maxwin, integer: true})
 
-                bot.chat.send(message.channel, {body: `The heist was *won*. ${players,join(', ')} ran off with *$${reward}* each!`})
+                bot.chat.send(message.channel, {body: `The heist was *won*. ${players.join(', ')} ran off with *$${reward}* each!`})
             } else {
                 reward = rn({min: settings.actions.heist.maxloss, max: settings.actions.heist.minloss, integer: true})
 
@@ -52,7 +52,7 @@ exports.run = (message, bot) => {
         timeout.refresh()
         winrate += playervalue
         players.push(message.sender.username)
-        bot.chat.send(message.channel, {body: `@${message.sender.username} joined the heist. *Win-rate increased to ${parseInt(winrate * 100)}%. Heist members: ${players.join(', ')}.`})
+        bot.chat.send(message.channel, {body: `@${message.sender.username} joined the heist. *Win-rate increased to ${parseInt(winrate * 100)}%. Timer reset to ${timer / 1000} seconds. Heist members: ${players.join(', ')}.`})
     }
 }