+++ commands/test.js
@@ -1,0 +1,4 @@
+
+exports.run = (message, bot, args) => {
+    console.log(message)
+}