+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+# 2/5/2020
+- Changed `+heist` to `+quest`!
+- Added a collaborative action `+heist`!
+- Added `+decode`!
+- Forced bot into `@underforums#gameroom` to promote https://underforums.com and keep other channels less spammy.
+
 # Initial Release 2/4/2020
 - Core features are (mostly) working. :)